+++ CHANGES.md
@@ -20,6 +20,7 @@ Changes in 11.7
 - Add "No Grades found for Percent" error in InputFinalGrades.php
 - Add RolloverDeleteCoursesSQL(), RolloverDoWarning(), RolloverUpdateDefaultSyearWarning() & RolloverUpdateDefaultSyear() functions in Rollover.fnc.php
 - Use RolloverDeleteCoursesSQL(), RolloverDoWarning(), RolloverUpdateDefaultSyearWarning() & RolloverUpdateDefaultSyear() functions in Rollover.php & Portal.php
+- Add & use RolloverAllSchoolsDoneWarning() function in Rollover.fnc.php & Rollover.php
 - Update Slovenian translation (76% complete), thanks to AT group in locale/sl_SI.utf8/
 - CSS adjust vertical alignment of buttons inside note/error message in stylesheet.css